+++ docs/source/conf.py
@@ -14,10 +14,9 @@ release = "0.2.0"
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ["autoapi.extension", "sphinxcontrib.programoutput"]
+extensions = ["autoapi.extension", "sphinxcontrib.programoutput", "myst_parser"]
 
 templates_path = ["_templates"]
-exclude_patterns = []
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -27,3 +26,10 @@ html_theme = "alabaster"
 html_static_path = ["build/html/_static"]
 
 autoapi_dirs = ["../../src/aoc_mod"]
+autoapi_ignore = ["*/templates/*"]
+
+source_suffix = {
+    '.rst': 'restructuredtext',
+    '.txt': 'markdown',
+    '.md': 'markdown'
+}